これまで反映していなかった内容を変更
