+++ osect_sensor/Application/edge_tcpdump/common/app_config.py
@@ -2,7 +2,7 @@ SLEEP_TIME = 60
 """ 監視インターバル """
 
 TCPDUMP_SHELL_COMMAND = ['/usr/sbin/tcpdump', '-w', 'realtime-%F-%T.pcap', '-G', '60', '-ni',
-                         'enp0s3', '-s', '0', '-z', '/opt/ot_tools/capture.sh']
+                         'enp0s3', '-s', '0', '-Z', 'root', '-z', '/opt/ot_tools/capture.sh']
 """ tcpdumpの起動shellコマンド """
 
 REMOVE_PCAP_FILE = 'realtime-*.pcap'
