+++ osect_sensor/Infrastructure/edge_cron/work/ot_tools/bro.sh
@@ -3,7 +3,7 @@
 cd $1/$2
 mkdir tmp
 cd tmp
-/opt/zeek/bin/zeek /opt/zeek/share/zeek/base/protocols/arp.zeek /opt/zeek/share/zeek/base/protocols/ns.zeek -r $1/$3
+/opt/zeek/bin/zeek /opt/zeek/share/zeek/base/protocols/arp.zeek /opt/zeek/share/zeek/base/protocols/ns.zeek /opt/zeek/share/zeek/base/protocols/consts_bacnet.zeek /opt/zeek/share/zeek/base/protocols/main_bacnet.zeek -r $1/$3
 mv {conn,arp,ns,dns,http}.log ../
 cd $1/$2
 rm -rf tmp
