+++ osect_sensor/Infrastructure/edge_cron/work/requirements.txt
@@ -1,18 +1,26 @@
+asgiref==3.4.1
 cachetools==4.2.2
 certifi==2020.6.20
-#certifi==2021.5.30
-chardet==3.0.4
-Django==3.2.5
+charset-normalizer==2.0.9
+Django==3.2.12
 dpkt==1.9.2
+gitdb2==2.0.3
+gitpython==3.0.7
 google-auth==1.30.0
 idna==2.10
+meson==0.56.0
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
 pytz==2019.1
-#pytz==2021.1
-requests==2.25.1
+PyYAML==6.0
+requests==2.26.0
 rsa==4.7.2
+semantic-version==2.8.2
 six==1.12.0
-#six==1.16.0
+smmap2==2.0.3
 sqlparse==0.4.2
+tornado==6.1
+typing_extensions==4.0.1
 urllib3==1.26.6
+wheel==0.34.2
+zkg==2.12.0
