+++ osect_sensor/Infrastructure/edge_tcpdump/work/requirements.txt
@@ -1,3 +1,5 @@
-Django==3.2.5
+asgiref==3.4.1
+Django==3.2.12
 pytz==2021.1
 sqlparse==0.4.2
+typing_extensions==4.0.1
